chore(restructure): split handlers into separate files
